// Back/src/shared/security/role.enum.ts
export enum Role {
    USER = "user",
    ADMIN = "admin",
    RECETA_T = "receta_admin",
    RECETA_L = "receta_lectura",
    RECETA_E = "receta_escritura",
    RECETA_D = "receta_borrado",
    INGREDIENTE_T = "ingrediente_admin",
    INGREDIENTE_L = "ingrediente_lectura",
    INGREDIENTE_E = "ingrediente_escritura",
    INGREDIENTE_D = "ingrediente_borrado",
    CANTIDAD_T = "cantidad_admin",
    CANTIDAD_L = "cantidad_lectura",
    CANTIDAD_E = "cantidad_escritura",
    CANTIDAD_D = "cantidad_borrado",
    RUTINA_RECETA_T = "rutina_receta_admin",
    RUTINA_RECETA_L = "rutina_receta_lectura",
    RUTINA_RECETA_E = "rutina_receta_escritura",
    RUTINA_RECETA_D = "rutina_receta_borrado",
    EJERCICIO_T="ejercicio_admin",
    EJERCICIO_L= "ejercicio__lectura",
    EJERCICIO_E= "ejercicio__escritura",
    EJERCICIO_D="ejercicio__borrado",
    RUTINA_T="rutina_admin",
    RUTINA_L= "rutina_lectura",
    RUTINA_E= "rutina_escritura",
    RUTINA_D="rutina_borrado",
    RUTINA_EJERCICIO_T= "rutina_ejercicio_admin",
    RUTINA_EJERCICIO_L= "rutina_ejercicio_lectura",
    RUTINA_EJERCICIO_E= "rutina_ejercicio_escritura",
    RUTINA_EJERCICIO_D= "rutina_ejercicio_borrado",
    RUTINA_PERFIL_T= "rutina_perfil_admin",
    RUTINA_PERFIL_L= "rutina_perfil_lectura",
    RUTINA_PERFIL_E= "rutina_perfil_escritura",
    RUTINA_PERFIL_D= "rutina_perfil_borrado",



    PERFIL_T = "perfil_admin",
    PERFIL_L = "perfil_lectura",
    PERFIL_E = "perfil_escritura",
    PERFIL_D = "perfil_borrado",
    METRICA_T = "metrica_admin",
    METRICA_L = "metrica_lectura",
    METRICA_E = "metrica_escritura",
    METRICA_D = "metrica_borrado",
    PERFIL_METRICA_T = "perfil_metrica_admin",
    PERFIL_METRICA_L = "perfil_metrica_lectura",
    PERFIL_METRICA_E = "perfil_metrica_escritura",
    PERFIL_METRICA_D = "perfil_metrica_borrado",
    PERFIL_LIKES_T = "perfil_likes_admin",
    PERFIL_LIKES_L = "perfil_likes_lectura",
    PERFIL_LIKES_E = "perfil_likes_escritura",
    PERFIL_LIKES_D = "perfil_likes_borrado",
    PERFIL_PUBLICACION_T = "perfil_publicacion_admin",
    PERFIL_PUBLICACION_L = "perfil_publicacion_lectura",
    PERFIL_PUBLICACION_E = "perfil_publicacion_escritura",
    PERFIL_PUBLICACION_D = "perfil_publicacion_borrado",
    PUBLICACION_T = "publicacion_admin",
    PUBLICACION_L = "publicacion_lectura",
    PUBLICACION_E = "publicacion_escritura",
    PUBLICACION_D = "publicacion_borrado",
    FORO_T = "foro_admin",
    FORO_L = "foro_lectura",
    FORO_E = "foro_escritura",
    FORO_D = "foro_borrado",
    FORO_PUBLICACION_D = "foro_publicacion_admin",
    FORO_PUBLICACION_E = "foro_publicacion_lectura",
    FORO_PUBLICACION_L = "foro_publicacion_escritura",
    FORO_PUBLICACION_T = "foro_publicacion_borrado"
}